Fix 'documentation', hide use of ash library from renderer interface
In order to have a place to put project-level documentation, `cargo doc` requires that there be a main package. Attempts at using a dummy package failed. As a consequence, all application entry points will have to be through the main package; not necessarily an issue.

Additionally, `RendererError` was leaking `ash::vk::Result`, which was fixed by adding a `VulkanError` newtype with appropriate conversions.