Test abstracting out pipeline stuff to an Effect
An Effect is basically a set of render pasees needed to apply some effect to a framebuffer. The current implementation is only half-baked (if that), but the idea is that other effects could be implemented in the same way. It is my hope that doing this will allow swapchains to keep a list of effects that were applied for a given frame, allowing unused ones to be destroyed, and new ones to be created.

There is some concern that the apply() method for every effect will be virtual, it is my hope and expectation that it won't be called more than a few hundred times per frame.

Next steps: implement swapchain resizing... yay...