Fix access violation on `vkCreateSubpass` in release mode
For whatever reason, the attachment reference was being constructed inline, which seemed to somehow cause it to be filled with garbage data, which then produced a crash when calling `vkCreateSubpass`.

Might be a compiler error here, but I have no idea how to minimize it to submit as a report.